ServiceWorker.ss Update
If the service worker gets turned off for any reason the notification url just defaults to "/".
This means that when any notificaiton is clicked it's  using the default notificationUrl = "/";
To fix this I made it so the notifications contain the url in the empty data field and pass it to the receiving notificationclick function.
The app now opens as expected and goes to the correct URL.

Took a loooooong time to get this one sorted. There is no example I could find on the internet replicating this issue.